pyproject.toml
[tool.poetry]
name = "neuromancer"
version = "0.2.1"
description = ""
authors = ["Sami Jawhar <sami@thecybermonk.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.10.0"

click = "^8.0.0"
dvc = "^2.9.4"
gitpython = "^3.1.24"
pyyaml = ">=5.4.0"

[tool.poetry.group.dev.dependencies]
black = "22.3.0"
pytest = "^7.1.2"
pytest-mock = "^3.7.0"
pytest-watch = "^4.2.0"

[tool.poetry.group.wintermute]
optional = true

[tool.poetry.group.wintermute.dependencies]
kubernetes = "^23.6.0"
dvc = { version = "^2.9.4", extras = ["s3", "ssh"] }


[tool.poetry.scripts]
neuromancer = "neuromancer.cli:neuromancer"
wintermute = "neuromancer.cli:wintermute"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py38"]

[tool.isort]
profile = "black"
known_local_folder = "neuromancer"
